+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.0.18] - 2020-09-25
+### Added
+- Add BINLOG_XXX in `PlatformConnectionGroup`
+- Add createBinlogXXX in `PlatformBaseModel`
 
 ## [1.0.17] - 2020-08-27
 ### Added